@@ -135,115 +135,103 @@ class _ImagePreviewState extends State<ImagePreview>
       ),
       body: Stack(
         children: [
-          DismissiblePage(
-            backgroundColor: Colors.transparent,
-            onDismissed: () {
-              Navigator.of(context).pop();
-            },
-            // Note that scrollable widget inside DismissiblePage might limit the functionality
-            // If scroll direction matches DismissiblePage direction
-            direction: DismissiblePageDismissDirection.down,
-            disabled: _dismissDisabled,
-            isFullScreen: true,
-            child: GestureDetector(
-              onLongPress: () => onOpenMenu(),
-              child: ExtendedImageGesturePageView.builder(
-                controller: ExtendedPageController(
-                  initialPage: _previewController.initialPage.value,
-                  pageSpacing: 0,
-                ),
-                onPageChanged: (int index) =>
-                    _previewController.onChange(index),
-                canScrollPage: (GestureDetails? gestureDetails) =>
-                    gestureDetails!.totalScale! <= 1.0,
-                itemCount: widget.imgList!.length,
-                itemBuilder: (BuildContext context, int index) {
-                  return ExtendedImage.network(
-                    widget.imgList![index],
-                    fit: BoxFit.contain,
-                    mode: ExtendedImageMode.gesture,
-                    onDoubleTap: (ExtendedImageGestureState state) {
-                      final Offset? pointerDownPosition =
-                          state.pointerDownPosition;
-                      final double? begin = state.gestureDetails!.totalScale;
-                      double end;
-
-                      //remove old
-                      _doubleClickAnimation
-                          ?.removeListener(_doubleClickAnimationListener);
-
-                      //stop pre
-                      _doubleClickAnimationController.stop();
-
-                      //reset to use
-                      _doubleClickAnimationController.reset();
-
-                      if (begin == doubleTapScales[0]) {
-                        setState(() {
-                          _dismissDisabled = true;
-                        });
-                        end = doubleTapScales[1];
-                      } else {
-                        setState(() {
-                          _dismissDisabled = false;
-                        });
-                        end = doubleTapScales[0];
-                      }
-
-                      _doubleClickAnimationListener = () {
-                        state.handleDoubleTap(
-                            scale: _doubleClickAnimation!.value,
-                            doubleTapPosition: pointerDownPosition);
-                      };
-                      _doubleClickAnimation = _doubleClickAnimationController
-                          .drive(Tween<double>(begin: begin, end: end));
-
-                      _doubleClickAnimation!
-                          .addListener(_doubleClickAnimationListener);
-
-                      _doubleClickAnimationController.forward();
-                    },
-                    // ignore: body_might_complete_normally_nullable
-                    loadStateChanged: (ExtendedImageState state) {
-                      if (state.extendedImageLoadState == LoadState.loading) {
-                        final ImageChunkEvent? loadingProgress =
-                            state.loadingProgress;
-                        final double? progress =
-                            loadingProgress?.expectedTotalBytes != null
-                                ? loadingProgress!.cumulativeBytesLoaded /
-                                    loadingProgress.expectedTotalBytes!
-                                : null;
-                        return Center(
-                          child: Column(
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            crossAxisAlignment: CrossAxisAlignment.center,
-                            children: <Widget>[
-                              SizedBox(
-                                width: 150.0,
-                                child: LinearProgressIndicator(
-                                  value: progress,
-                                  color: Colors.white,
-                                ),
-                              ),
-                              // const SizedBox(height: 10.0),
-                              // Text('${((progress ?? 0.0) * 100).toInt()}%',),
-                            ],
-                          ),
-                        );
-                      }
-                    },
-                    initGestureConfigHandler: (ExtendedImageState state) {
-                      return GestureConfig(
-                        inPageView: true,
-                        initialScale: 1.0,
-                        maxScale: 5.0,
-                        animationMaxScale: 6.0,
-                        initialAlignment: InitialAlignment.center,
-                      );
-                    },
-                  );
-                },
+          GestureDetector(
+            onLongPress: () => onOpenMenu(),
+            child: ExtendedImageGesturePageView.builder(
+              controller: ExtendedPageController(
+                initialPage: _previewController.initialPage.value,
+                pageSpacing: 0,
               ),
+              onPageChanged: (int index) => _previewController.onChange(index),
+              canScrollPage: (GestureDetails? gestureDetails) =>
+                  gestureDetails!.totalScale! <= 1.0,
+              itemCount: widget.imgList!.length,
+              itemBuilder: (BuildContext context, int index) {
+                return ExtendedImage.network(
+                  widget.imgList![index],
+                  fit: BoxFit.contain,
+                  mode: ExtendedImageMode.gesture,
+                  onDoubleTap: (ExtendedImageGestureState state) {
+                    final Offset? pointerDownPosition =
+                        state.pointerDownPosition;
+                    final double? begin = state.gestureDetails!.totalScale;
+                    double end;
+
+                    //remove old
+                    _doubleClickAnimation
+                        ?.removeListener(_doubleClickAnimationListener);
+
+                    //stop pre
+                    _doubleClickAnimationController.stop();
+
+                    //reset to use
+                    _doubleClickAnimationController.reset();
+
+                    if (begin == doubleTapScales[0]) {
+                      setState(() {
+                        _dismissDisabled = true;
+                      });
+                      end = doubleTapScales[1];
+                    } else {
+                      setState(() {
+                        _dismissDisabled = false;
+                      });
+                      end = doubleTapScales[0];
+                    }
+
+                    _doubleClickAnimationListener = () {
+                      state.handleDoubleTap(
+                          scale: _doubleClickAnimation!.value,
+                          doubleTapPosition: pointerDownPosition);
+                    };
+                    _doubleClickAnimation = _doubleClickAnimationController
+                        .drive(Tween<double>(begin: begin, end: end));
+
+                    _doubleClickAnimation!
+                        .addListener(_doubleClickAnimationListener);
+
+                    _doubleClickAnimationController.forward();
+                  },
+                  // ignore: body_might_complete_normally_nullable
+                  loadStateChanged: (ExtendedImageState state) {
+                    if (state.extendedImageLoadState == LoadState.loading) {
+                      final ImageChunkEvent? loadingProgress =
+                          state.loadingProgress;
+                      final double? progress =
+                          loadingProgress?.expectedTotalBytes != null
+                              ? loadingProgress!.cumulativeBytesLoaded /
+                                  loadingProgress.expectedTotalBytes!
+                              : null;
+                      return Center(
+                        child: Column(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          crossAxisAlignment: CrossAxisAlignment.center,
+                          children: <Widget>[
+                            SizedBox(
+                              width: 150.0,
+                              child: LinearProgressIndicator(
+                                value: progress,
+                                color: Colors.white,
+                              ),
+                            ),
+                            // const SizedBox(height: 10.0),
+                            // Text('${((progress ?? 0.0) * 100).toInt()}%',),
+                          ],
+                        ),
+                      );
+                    }
+                  },
+                  initGestureConfigHandler: (ExtendedImageState state) {
+                    return GestureConfig(
+                      inPageView: true,
+                      initialScale: 1.0,
+                      maxScale: 5.0,
+                      animationMaxScale: 6.0,
+                      initialAlignment: InitialAlignment.center,
+                    );
+                  },
+                );
+              },
             ),
           ),
           Positioned(
@@ -251,33 +239,49 @@ class _ImagePreviewState extends State<ImagePreview>
             right: 0,
             bottom: 0,
             child: Container(
-              padding: EdgeInsets.only(
-                  bottom: MediaQuery.of(context).padding.bottom + 30),
-              decoration: const BoxDecoration(
-                gradient: LinearGradient(
-                  begin: Alignment.topCenter,
-                  end: Alignment.bottomCenter,
-                  colors: <Color>[
-                    Colors.transparent,
-                    Colors.black87,
+                padding: EdgeInsets.only(
+                    left: 20,
+                    right: 20,
+                    bottom: MediaQuery.of(context).padding.bottom + 30),
+                decoration: const BoxDecoration(
+                  gradient: LinearGradient(
+                    begin: Alignment.topCenter,
+                    end: Alignment.bottomCenter,
+                    colors: <Color>[
+                      Colors.transparent,
+                      Colors.black87,
+                    ],
+                    tileMode: TileMode.mirror,
+                  ),
+                ),
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                  children: [
+                    widget.imgList!.length > 1
+                        ? Obx(
+                            () => Text.rich(
+                              textAlign: TextAlign.center,
+                              TextSpan(
+                                  style: const TextStyle(
+                                      color: Colors.white, fontSize: 16),
+                                  children: [
+                                    TextSpan(
+                                        text: _previewController.currentPage
+                                            .toString()),
+                                    const TextSpan(text: ' / '),
+                                    TextSpan(
+                                        text:
+                                            widget.imgList!.length.toString()),
+                                  ]),
+                            ),
+                          )
+                        : const SizedBox(),
+                    IconButton(
+                      onPressed: () => Get.back(),
+                      icon: const Icon(Icons.close, color: Colors.white),
+                    ),
                   ],
-                  tileMode: TileMode.mirror,
-                ),
-              ),
-              child: Obx(
-                () => Text.rich(
-                  textAlign: TextAlign.center,
-                  TextSpan(
-                      style: const TextStyle(color: Colors.white, fontSize: 15),
-                      children: [
-                        TextSpan(
-                            text: _previewController.currentPage.toString()),
-                        const TextSpan(text: ' / '),
-                        TextSpan(text: widget.imgList!.length.toString()),
-                      ]),
-                ),
-              ),
-            ),
+                )),
           ),
         ],
       ),
